V "1.44"
In this commit, I have implemented a caching mechanism for our list views, which is now controlled by signals. This means that whenever an object is added or deleted in the database, the cache is automatically updated - removing the outdated data and creating a new cache to reflect the changes.